feat: 좌석 선택 후 예매 시 다이얼로그 출력
