+++ app/src/main/java/woowacourse/movie/view/reservation/SeatSelectContract.kt
@@ -1,0 +1,12 @@
+package woowacourse.movie.view.reservation
+
+interface SeatSelectContract {
+    interface View {
+        fun showReservationDialog(
+            title: String,
+            message: String,
+        )
+    }
+
+    interface Presenter
+}
